refactor(stream): separate private media and codec helpers
## What

- Organize private media matching and request codec-finalization helpers into focused source files without changing their behavior.

## Why

- Keep related implementation concerns localized so future maintenance remains clear while preserving existing HTTP streaming behavior.

## Testing

- `make package=net/http/content/stream specs` - passed
- Existing package coverage exercises media selection and request codec-close behavior; no behavior change required new tests.
- `make package=net/http/content/stream lint` - passed

AI-assisted-by: [Codex](https://openai.com/codex/)
